requirements.txt
pandas>=1.3.0
google-analytics-data>=0.14.0
matplotlib>=3.5.0  # Optional, for visualization
